Updating FMS-moving out of figure file and plotting functions
